refactor(scripts): reorganize TypeScript modules into feature-based structure
- Split monolithic utilities.ts into dedicated modules:
  - core/windowmanager.ts: Window management (focus, drag, minimize/maximize)
  - features/processmonitor.ts: htop-style process monitor
  - features/mouse-settings.ts: Mouse panel configuration
  - utils/clock.ts: System clock functionality
  - utils/beep.ts: Audio beep utility
  - utils/screenshot.ts: Screenshot capture with html2canvas

- Create boot/init.ts as central initialization entry point
- Update imports across all modules to reflect new structure
- Add proper TypeScript interfaces and JSDoc documentation
- Implement professional logging with [Module] prefixes
- Fix window dragging by adding automatic titlebar initialization
- Ensure all modules self-register globals for HTML compatibility
- Maintain existing functionality while improving code organization

This modularization improves maintainability, testability, and
separation of concerns across the entire codebase.
