+++ src/components/apps/FileManager.astro
@@ -68,7 +68,7 @@ import WindowControls from "../common/WindowControls.astro";
 </div>
 
 <script>
-  import('../../scripts/filemanager.ts').then((module) => {
+  import('../../scripts/features/filemanager.ts').then((module) => {
     module.FileManager.init();
   }).catch(error => {
     console.error('Error al cargar FileManager:', error);
