+++ src/layouts/layout.astro
@@ -21,7 +21,6 @@ const { title } = Astro.props;
     <link rel="stylesheet" href="/css/core.css" />
     <link rel="stylesheet" href="/css/components.css" />
     <link rel="stylesheet" href="/css/app.css" />
-    <link rel="stylesheet" href="/css/effects.css" />
   </head>
   <body>
     <slot />
