+++ src/scripts/features/filemanager.ts
@@ -1,64 +1,137 @@
 // src/scripts/filemanager.ts
 
-import { CONFIG } from './config';
-import filesystemData from '../data/filesystem.json';
-import { CDEModal } from './modals';
+import { CONFIG } from '../core/config';
+import filesystemData from '../../data/filesystem.json';
+import { CDEModal } from '../ui/modals';
+
 declare global {
   interface Window {
+    /**
+     * Opens the file manager window.
+     */
     openFileManager: () => void;
+    
+    /**
+     * Closes the file manager window.
+     */
     closeFileManager: () => void;
+    
+    /**
+     * Toggles the file manager window visibility.
+     */
     toggleFileManager: () => void;
+    
+    /**
+     * Checks if the file manager is currently open.
+     * @returns {boolean} True if the file manager is open, false otherwise.
+     */
     isFileManagerOpen: () => boolean;
+    
+    /**
+     * Navigates to the specified path in the file manager.
+     * @param path - The path to navigate to.
+     */
     openPath: (path: string) => void;
+    
+    /**
+     * Navigates back in the file manager history.
+     */
     goBack: () => void;
+    
+    /**
+     * Navigates forward in the file manager history.
+     */
     goForward: () => void;
+    
+    /**
+     * Navigates up one directory level.
+     */
     goUp: () => void;
+    
+    /**
+     * Navigates to the home directory.
+     */
     goHome: () => void;
   }
 }
 
 // ------------------------------------------------------------------
-// TIPOS DEL SISTEMA DE ARCHIVOS VIRTUAL
+// VIRTUAL FILESYSTEM TYPES
 // ------------------------------------------------------------------
+
+/** Represents a file in the virtual filesystem */
 type FileType = 'file' | 'folder';
 
+/** Interface for a virtual file */
 interface VirtualFile {
   type: 'file';
+  /** File content as string */
   content: string;
 }
 
+/** Interface for a virtual folder */
 interface VirtualFolder {
   type: 'folder';
+  /** Children nodes contained in this folder */
   children: Record<string, VirtualFileSystemNode>;
 }
 
+/** Union type for filesystem nodes (file or folder) */
 type VirtualFileSystemNode = VirtualFile | VirtualFolder;
 
+// ------------------------------------------------------------------
+// INTERNAL STATE
+// ------------------------------------------------------------------
+
+/** Virtual filesystem data loaded from JSON */
 const fs = filesystemData as Record<string, VirtualFolder>;
 
+/** Current working directory path */
 let currentPath: string = CONFIG.FS.HOME;
+
+/** Navigation history stack */
 let history: string[] = [];
+
+/** Current index in navigation history */
 let historyIndex: number = -1;
+
+/** Currently selected file/folder name */
 let fmSelected: string | null = null;
+
+/** Whether to show hidden files (starting with dot) */
 let showHidden: boolean = false;
+
+/** Z-index counter for window layering */
 let zIndex: number = CONFIG.FILEMANAGER.BASE_Z_INDEX;
+
+/** Whether the file manager has been initialized */
 let initialized: boolean = false;
+
+/** Reference to the currently active dropdown menu */
 let activeMenu: HTMLElement | null = null;
+
+/** Reference to the currently active context menu */
 let activeContextMenu: HTMLElement | null = null;
 
 // ------------------------------------------------------------------
-// FUNCIONES PRIVADAS
+// PRIVATE FUNCTIONS
 // ------------------------------------------------------------------
 
 /**
- * Obtiene el contenido de la carpeta actual.
+ * Retrieves the contents of the current folder.
+ * 
+ * @returns {Record<string, VirtualFileSystemNode> | null} The folder contents or null if not found
  */
 function getCurrentFolder(): Record<string, VirtualFileSystemNode> | null {
   return fs[currentPath]?.children || null;
 }
 
 /**
- * Renderiza los archivos en la interfaz.
+ * Renders the files in the current folder to the UI.
+ * 
+ * @remarks
+ * Updates the file list display, path input, and status bar.
+ * Handles hidden file filtering based on showHidden flag.
  */
 function renderFiles(): void {
   const container = document.getElementById('fmFiles');
@@ -66,7 +139,7 @@ function renderFiles(): void {
   const status = document.getElementById('fmStatus');
 
   if (!container || !pathInput || !status) {
-    console.warn('File Manager: elementos no encontrados');
+    console.warn('[FileManager] renderFiles: required UI elements not found');
     return;
   }
 
@@ -75,7 +148,7 @@ function renderFiles(): void {
 
   const folder = getCurrentFolder();
   if (!folder) {
-    console.warn(`File Manager: ruta no encontrada: ${currentPath}`);
+    console.warn(`[FileManager] renderFiles: path not found: ${currentPath}`);
     return;
   }
 
@@ -117,112 +190,184 @@ function renderFiles(): void {
   });
 
   status.textContent = `${count} ${count === 1 ? 'item' : 'items'}`;
+  console.log(`[FileManager] renderFiles: displayed ${count} items at path ${currentPath}`);
 }
 
 /**
- * Abre una ruta del sistema de archivos.
+ * Navigates to the specified filesystem path.
+ * 
+ * @param path - The path to navigate to
+ * 
+ * @remarks
+ * Updates navigation history and current path, then re-renders.
+ * Does nothing if the path does not exist in the filesystem.
  */
 function openPath(path: string): void {
-  if (!fs[path]) return;
+  if (!fs[path]) {
+    console.warn(`[FileManager] openPath: path does not exist: ${path}`);
+    return;
+  }
+  
   history = history.slice(0, historyIndex + 1);
   history.push(path);
   historyIndex++;
   currentPath = path;
+  
+  console.log(`[FileManager] openPath: navigated to ${path}`);
   renderFiles();
 }
 
 /**
- * Navega hacia atrás en el historial.
+ * Navigates back in the navigation history.
  */
 function goBack(): void {
   if (historyIndex > 0) {
     historyIndex--;
     currentPath = history[historyIndex];
+    console.log(`[FileManager] goBack: navigated to ${currentPath}`);
     renderFiles();
+  } else {
+    console.log('[FileManager] goBack: no previous history entry');
   }
 }
 
 /**
- * Navega hacia adelante en el historial.
+ * Navigates forward in the navigation history.
  */
 function goForward(): void {
   if (historyIndex < history.length - 1) {
     historyIndex++;
     currentPath = history[historyIndex];
+    console.log(`[FileManager] goForward: navigated to ${currentPath}`);
     renderFiles();
+  } else {
+    console.log('[FileManager] goForward: no forward history entry');
   }
 }
 
 /**
- * Navega al directorio padre.
+ * Navigates to the parent directory.
  */
 function goUp(): void {
   const parent = currentPath.split('/').slice(0, -2).join('/') + '/';
-  if (fs[parent]) openPath(parent);
+  if (fs[parent]) {
+    console.log(`[FileManager] goUp: from ${currentPath} to ${parent}`);
+    openPath(parent);
+  } else {
+    console.log(`[FileManager] goUp: no parent directory for ${currentPath}`);
+  }
 }
 
 /**
- * Navega al directorio home.
+ * Navigates to the home directory.
  */
 function goHome(): void {
+  console.log('[FileManager] goHome: navigating to home directory');
   openPath(CONFIG.FS.HOME);
 }
 
 // ------------------------------------------------------------------
-// OPERACIONES ASÍNCRONAS CON MODALES
+// ASYNCHRONOUS OPERATIONS WITH MODALS
 // ------------------------------------------------------------------
 
+/**
+ * Creates a new empty file with the specified name.
+ * 
+ * @param name - The name of the file to create
+ * @returns Promise that resolves when the file is created
+ */
 async function touch(name: string): Promise<void> {
   const folder = getCurrentFolder();
   if (folder && name) {
     folder[name] = { type: 'file', content: '' } as VirtualFile;
+    console.log(`[FileManager] touch: created file "${name}"`);
     renderFiles();
   }
 }
 
+/**
+ * Creates a new folder with the specified name.
+ * 
+ * @param name - The name of the folder to create
+ * @returns Promise that resolves when the folder is created
+ */
 async function mkdir(name: string): Promise<void> {
   const folder = getCurrentFolder();
   if (folder && name) {
     folder[name] = { type: 'folder', children: {} } as VirtualFolder;
+    console.log(`[FileManager] mkdir: created folder "${name}"`);
     renderFiles();
   }
 }
 
+/**
+ * Deletes the specified file or folder.
+ * 
+ * @param name - The name of the item to delete
+ * @returns Promise that resolves when the deletion is complete
+ * 
+ * @remarks
+ * Shows a confirmation modal before deletion.
+ */
 async function rm(name: string): Promise<void> {
   const folder = getCurrentFolder();
   if (!folder) return;
+  
   const confirmed = await CDEModal.confirm(`Delete ${name}?`);
   if (confirmed) {
     delete folder[name];
     fmSelected = null;
+    console.log(`[FileManager] rm: deleted "${name}"`);
     renderFiles();
   }
 }
 
+/**
+ * Renames a file or folder.
+ * 
+ * @param oldName - The current name of the item
+ * @param newName - The new name for the item
+ * @returns Promise that resolves when the rename is complete
+ */
 async function rename(oldName: string, newName: string): Promise<void> {
   const folder = getCurrentFolder();
   if (folder && folder[oldName] && newName && newName !== oldName) {
     folder[newName] = folder[oldName];
     delete folder[oldName];
     fmSelected = null;
+    console.log(`[FileManager] rename: renamed "${oldName}" to "${newName}"`);
     renderFiles();
   }
 }
 
+/**
+ * Opens a text file in a modal window.
+ * 
+ * @param name - The name of the file
+ * @param content - The content of the file to display
+ */
 function openTextWindow(name: string, content: string): void {
   const win = document.getElementById('textWindow') as HTMLElement | null;
   if (!win) {
-    console.warn('File Manager: elemento #textWindow no encontrado');
+    console.warn('[FileManager] openTextWindow: #textWindow element not found');
     return;
   }
+  
   const titleEl = document.getElementById('textTitle') as HTMLElement | null;
   const contentEl = document.getElementById('textContent') as HTMLElement | null;
+  
   if (titleEl) titleEl.textContent = name;
   if (contentEl) contentEl.textContent = content;
+  
   win.style.display = 'block';
   win.style.zIndex = String(++zIndex);
+  
+  console.log(`[FileManager] openTextWindow: opened file "${name}"`);
 }
 
+/**
+ * Closes the currently active dropdown menu.
+ */
 function closeMenu(): void {
   if (activeMenu) {
     activeMenu.remove();
@@ -230,6 +375,9 @@ function closeMenu(): void {
   }
 }
 
+/**
+ * Closes the currently active context menu.
+ */
 function closeContextMenu(): void {
   if (activeContextMenu) {
     activeContextMenu.remove();
@@ -238,14 +386,18 @@ function closeContextMenu(): void {
 }
 
 // ------------------------------------------------------------------
-// MENÚS (INTERFAZ DE USUARIO EN INGLÉS)
+// MENUS (UI IN ENGLISH)
 // ------------------------------------------------------------------
 
+/** Represents a menu item in the file manager menu bar */
 interface MenuItem {
+  /** The display label for the menu item */
   label: string;
+  /** The action to execute when the menu item is clicked */
   action: () => Promise<void> | void;
 }
 
+/** Menu definitions for the file manager menu bar */
 const fmMenus: Record<string, MenuItem[]> = {
   File: [
     {
@@ -284,12 +436,16 @@ const fmMenus: Record<string, MenuItem[]> = {
       label: 'Show Hidden Files',
       action: () => {
         showHidden = !showHidden;
+        console.log(`[FileManager] showHidden: toggled to ${showHidden}`);
         renderFiles();
       },
     },
     {
       label: 'Refresh',
-      action: () => renderFiles(),
+      action: () => {
+        console.log('[FileManager] refresh: re-rendering files');
+        renderFiles();
+      },
     },
   ],
   Go: [
@@ -300,11 +456,17 @@ const fmMenus: Record<string, MenuItem[]> = {
   ],
 };
 
+/**
+ * Sets up the menu bar event listeners.
+ */
 function setupMenuBar(): void {
   const menuBar = document.querySelector('.fm-menubar');
-  if (!menuBar) return;
+  if (!menuBar) {
+    console.warn('[FileManager] setupMenuBar: menu bar element not found');
+    return;
+  }
 
-  // Eliminar listeners anteriores clonando y reemplazando
+  // Remove previous listeners by cloning and replacing
   menuBar.querySelectorAll('span').forEach((span) => {
     const newSpan = span.cloneNode(true) as HTMLSpanElement;
     span.parentNode?.replaceChild(newSpan, span);
@@ -318,7 +480,10 @@ function setupMenuBar(): void {
 
       const name = span.textContent?.trim() || '';
       const items = fmMenus[name];
-      if (!items) return;
+      if (!items) {
+        console.log(`[FileManager] setupMenuBar: no items for menu "${name}"`);
+        return;
+      }
 
       const menu = document.createElement('div');
       menu.className = 'fm-dropdown';
@@ -332,7 +497,7 @@ function setupMenuBar(): void {
           try {
             await item.action();
           } catch (error) {
-            console.error('Error en acción del menú:', error);
+            console.error('[FileManager] menu action error:', error);
           }
           closeMenu();
         });
@@ -344,18 +509,35 @@ function setupMenuBar(): void {
       menu.style.left = rect.left + 'px';
       menu.style.top = rect.bottom + 'px';
       activeMenu = menu;
+      
+      console.log(`[FileManager] setupMenuBar: opened "${name}" menu`);
     });
   });
+  
+  console.log('[FileManager] setupMenuBar: menu bar initialized');
 }
 
+/**
+ * Sets up the context menu for file items.
+ */
 function setupContextMenu(): void {
   const fmFiles = document.getElementById('fmFiles');
-  if (!fmFiles) return;
+  if (!fmFiles) {
+    console.warn('[FileManager] setupContextMenu: fmFiles element not found');
+    return;
+  }
 
   fmFiles.removeEventListener('contextmenu', handleContextMenu);
   fmFiles.addEventListener('contextmenu', handleContextMenu);
+  
+  console.log('[FileManager] setupContextMenu: context menu initialized');
 }
 
+/**
+ * Handles right-click context menu events on file items.
+ * 
+ * @param e - The mouse event that triggered the context menu
+ */
 async function handleContextMenu(e: MouseEvent): Promise<void> {
   e.preventDefault();
   closeContextMenu();
@@ -384,8 +566,11 @@ async function handleContextMenu(e: MouseEvent): Promise<void> {
         if (!folder) return;
         const item = folder[name];
         if (item) {
-          if (item.type === 'folder') openPath(currentPath + name + '/');
-          else openTextWindow(name, (item as VirtualFile).content);
+          if (item.type === 'folder') {
+            openPath(currentPath + name + '/');
+          } else {
+            openTextWindow(name, (item as VirtualFile).content);
+          }
         }
       },
     },
@@ -421,7 +606,7 @@ async function handleContextMenu(e: MouseEvent): Promise<void> {
       try {
         await item.action();
       } catch (error) {
-        console.error('Error en menú contextual:', error);
+        console.error('[FileManager] context menu action error:', error);
       }
       closeContextMenu();
     });
@@ -432,21 +617,36 @@ async function handleContextMenu(e: MouseEvent): Promise<void> {
   menu.style.top = e.pageY + 'px';
   document.body.appendChild(menu);
 
-  // Ajustar posición si se sale de la pantalla
+  // Adjust position if menu goes off screen
   const menuRect = menu.getBoundingClientRect();
   const vw = window.innerWidth;
   const vh = window.innerHeight;
-  if (menuRect.right > vw) menu.style.left = e.pageX - menuRect.width + 'px';
-  if (menuRect.bottom > vh) menu.style.top = e.pageY - menuRect.height + 'px';
+  
+  if (menuRect.right > vw) {
+    menu.style.left = e.pageX - menuRect.width + 'px';
+  }
+  if (menuRect.bottom > vh) {
+    menu.style.top = e.pageY - menuRect.height + 'px';
+  }
 
   activeContextMenu = menu;
+  console.log(`[FileManager] handleContextMenu: opened context menu for "${name}"`);
 }
 
 // ------------------------------------------------------------------
-// API PÚBLICA (expuesta al ámbito global)
+// PUBLIC API (exposed to global scope)
 // ------------------------------------------------------------------
 
+/**
+ * Initializes the file manager.
+ * 
+ * @remarks
+ * Sets up the menu bar, context menu, sidebar navigation, and renders the initial files.
+ * Should be called once before using the file manager.
+ */
 function init(): void {
+  console.log('[FileManager] init: initializing file manager');
+  
   currentPath = CONFIG.FS.HOME;
   history = [currentPath];
   historyIndex = 0;
@@ -473,54 +673,115 @@ function init(): void {
   });
 
   initialized = true;
+  console.log('[FileManager] init: initialization complete');
 }
 
+/**
+ * Opens the file manager window.
+ */
 function open(): void {
   const win = document.getElementById('fm') as HTMLElement | null;
-  if (!win) return;
-  if (!initialized) init();
+  if (!win) {
+    console.warn('[FileManager] open: file manager window element not found');
+    return;
+  }
+  
+  if (!initialized) {
+    console.log('[FileManager] open: not initialized, initializing now');
+    init();
+  }
+  
   win.style.display = 'block';
   win.style.zIndex = String(++zIndex);
+  
+  console.log('[FileManager] open: file manager window opened');
   renderFiles();
 }
 
+/**
+ * Closes the file manager window.
+ */
 function close(): void {
   const win = document.getElementById('fm') as HTMLElement | null;
-  if (win) win.style.display = 'none';
+  if (win) {
+    win.style.display = 'none';
+    console.log('[FileManager] close: file manager window closed');
+  } else {
+    console.warn('[FileManager] close: file manager window element not found');
+  }
 }
 
+/**
+ * Toggles the file manager window visibility.
+ */
 function toggle(): void {
   const win = document.getElementById('fm') as HTMLElement | null;
-  if (!win) return;
-  if (!initialized) init();
+  if (!win) {
+    console.warn('[FileManager] toggle: file manager window element not found');
+    return;
+  }
+  
+  if (!initialized) {
+    console.log('[FileManager] toggle: not initialized, initializing now');
+    init();
+  }
+  
   if (win.style.display === 'none' || win.style.display === '') {
     win.style.display = 'block';
     win.style.zIndex = String(++zIndex);
+    console.log('[FileManager] toggle: file manager opened');
   } else {
     win.style.zIndex = String(++zIndex);
+    console.log('[FileManager] toggle: file manager brought to front');
   }
+  
   renderFiles();
 }
 
+/**
+ * Checks if the file manager is currently open.
+ * 
+ * @returns {boolean} True if the file manager is open, false otherwise
+ */
 function isOpen(): boolean {
   const win = document.getElementById('fm') as HTMLElement | null;
-  return win?.style.display !== 'none' && win?.style.display !== '';
+  const isVisible = win?.style.display !== 'none' && win?.style.display !== '';
+  console.log(`[FileManager] isOpen: ${isVisible}`);
+  return isVisible;
 }
 
+/**
+ * File Manager module for navigating and managing a virtual filesystem.
+ * 
+ * @remarks
+ * Provides a complete file manager interface with navigation history,
+ * context menus, file operations (create, delete, rename), and a modal
+ * system for user interactions.
+ */
 export const FileManager = {
+  /** Initializes the file manager */
   init,
+  /** Opens the file manager window */
   open,
+  /** Closes the file manager window */
   close,
+  /** Toggles the file manager window visibility */
   toggle,
+  /** Checks if the file manager is open */
   isOpen,
+  /** Navigates to the specified path */
   openPath,
+  /** Navigates back in history */
   goBack,
+  /** Navigates forward in history */
   goForward,
+  /** Navigates up one directory level */
   goUp,
+  /** Navigates to the home directory */
   goHome,
 };
 
-// Exponer al ámbito global
+// Expose to global scope
 window.openFileManager = open;
 window.closeFileManager = close;
 window.toggleFileManager = toggle;
